Refactor inbox as zapier
