--- config/services/zapier.js
@@ -1,5 +1,0 @@
-'use strict';
-
-module.exports = {
-  webhookUrl: process.env.ZAPIER_WEBHOOK_URL,
-};
